% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpsro-package.R
\docType{package}
\name{hpsro-package}
\alias{hpsro}
\alias{hpsro-package}
\title{hpsro: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChristopherNClark/hpsro}
  \item Report bugs at \url{https://github.com/ChristopherNClark/hpsro/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
