% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_week_number.R
\name{load_data_week_number}
\alias{load_data_week_number}
\title{Load the data by week number}
\usage{
load_data_week_number(week)
}
\arguments{
\item{week}{Week number}
}
\value{
data
}
\description{
Load the data by week number
}
\examples{
\dontrun{
Example 1: Import week 1 data:
load_data_week_number(1)
}
}
